convert to base and cub
