+++ src/pytorch_metric_learning/datasets/cub.py
@@ -1,38 +1,12 @@
-from PIL import Image
-from torch.utils.data import Dataset
+from ..datasets.base_dataset import BaseDataset
 import os
 
-class CUB(Dataset):
+class CUB(BaseDataset):
 
-    SPLITS = ["train", "test", "train+test"]
-    DOWNLOAD_URL = "https://data.caltech.edu/records/65de6-vp158/files/CUB_200_2011.tgz"
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
 
-    def __init__(self, root, split="train+test", transform=None, target_transform=None, download=False):
-        self.root = root
-
-        if download and not os.path.isdir(self.root):
-            archive_name = CUB.DOWNLOAD_URL.split('/')[-1]
-            os.makedirs(self.root, exist_ok=False)
-            os.system(f"wget -P {self.root} {CUB.DOWNLOAD_URL}")
-            os.system(f"cd {self.root} && tar -xzvf {archive_name}")
-            os.system(f"rm {os.path.join(self.root, archive_name)}")
-        else:
-            # The given directory does not exist so the user should be aware of downloading it
-            # Otherwise proceed as usual
-            if not os.path.isdir(self.root):
-                raise ValueError("The given path does not exist. "
-                    "You should probably initialize the dataset with download=True."
-                )
-
-        self.transform = transform
-        self.target_transform = target_transform
-
-        if split not in CUB.SPLITS:
-            raise ValueError(f"Supported splits are: {', '.join(CUB.SPLITS)}")
-        
-        self.split = split
-
-        dir_name = CUB.DOWNLOAD_URL.split('/')[-1].replace(".tgz", "")
+        dir_name = self.get_download_url().split('/')[-1].replace(".tgz", "")
 
         # Training split is first 100 classes, other 100 is test
         if self.split == "train":
@@ -48,6 +22,10 @@ class CUB(Dataset):
             with open(os.path.join(self.root, dir_name, "images.txt")) as f2:
                 for l1, l2 in zip(f1, f2):
                     img_idx1, class_idx = list(map(int, l1.split()))
+                    
+                    if class_idx not in classes:
+                        continue
+
                     img_idx2, img_path = l2.split()
                     img_idx2 = int(img_idx2)
 
@@ -61,17 +39,16 @@ class CUB(Dataset):
         # Normalize labels to start from 0
         self.labels = [x - min(self.labels) for x in self.labels]
 
-    def __len__(self):
-        return len(self.labels)
-    
-    def __getitem__(self, idx):
-        img = Image.open(self.paths[idx])
-        label = self.labels[idx]
+    def download_and_remove(self):
+        archive_name = self.get_download_url().split('/')[-1]
+        os.system(f"wget -P {self.root} {self.get_download_url()}")
+        os.system(f"cd {self.root} && tar -xzvf {archive_name}")
+        os.system(f"rm {os.path.join(self.root, archive_name)}")
 
-        if self.transform is not None:
-            img = self.transform(img)
+    @staticmethod
+    def get_available_splits():
+        return ["train", "test", "train+test"]
 
-        if self.target_transform is not None:
-            label = self.target_transform(label)
-
-        return (img, label)
+    @staticmethod
+    def get_download_url():
+        return "https://data.caltech.edu/records/65de6-vp158/files/CUB_200_2011.tgz"
